superstructure

+++ src/robot/sensor/RobotEncoder.h
@@ -1,4 +1,4 @@
-int readEncoder(bool leftSide);
+double readEncoder(bool leftSide);
 void init();
 void updateEncoder();
